Membuat Permission 01

+++ app/Http/Controllers/AcademicSemesterController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\AcademicSemester;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use App\Http\Requests\StoreAcademicSemesterRequest;
 use App\Http\Requests\UpdateAcademicSemesterRequest;
@@ -10,14 +11,18 @@ use App\Http\Requests\UpdateAcademicSemesterRequest;
 
 class AcademicSemesterController extends Controller
 {
+    use AuthorizesRequests;
+
     public function index()
     {
+        $this->authorize('view_academic-semesters');
         $academicSemesters = AcademicSemester::latest()->get();
         return view('dashboard.admin.academic-semesters.index', compact('academicSemesters'));
     }
 
     public function create()
     {
+        $this->authorize('create_academic-semesters');
         return view('dashboard.admin.academic-semesters.create');
     }
 
@@ -31,6 +36,7 @@ class AcademicSemesterController extends Controller
 
     public function edit(AcademicSemester $academicSemester)
     {
+        $this->authorize('edit_academic-semesters');
         return view('dashboard.admin.academic-semesters.edit', compact('academicSemester'));
     }
 
@@ -44,6 +50,7 @@ class AcademicSemesterController extends Controller
 
     public function destroy(AcademicSemester $academicSemester)
     {
+        $this->authorize('delete_academic-semesters');
         $academicSemester->delete();
 
         return redirect()->route('dashboard.academic-semesters.index')
