+++ app/Http/Controllers/CurriculumController.php
@@ -7,20 +7,25 @@ use App\Models\Course;
 use App\Models\StudyProgram;
 use App\Models\AcademicSemester;
 use App\Models\Semester;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use App\Http\Requests\StoreCurriculumRequest;
 use App\Http\Requests\UpdateCurriculumRequest;
 
 class CurriculumController extends Controller
 {
+    use AuthorizesRequests;
+
     public function index()
     {
+        $this->authorize('view_curriculums');
         $curriculums = Curriculum::with(['course', 'studyProgram', 'academicSemester', 'semester'])->get();
         return view('dashboard.admin.curriculums.index', compact('curriculums'));
     }
 
     public function create()
     {
+        $this->authorize('create_curriculums');
         $courses = Course::all();
         $studyPrograms = StudyProgram::all();
         $academicSemesters = AcademicSemester::all();
@@ -38,6 +43,7 @@ class CurriculumController extends Controller
 
     public function edit(Curriculum $curriculum)
     {
+        $this->authorize('edit_curriculums');
         $courses = Course::all();
         $studyPrograms = StudyProgram::all();
         $academicSemesters = AcademicSemester::all();
@@ -55,6 +61,7 @@ class CurriculumController extends Controller
 
     public function destroy(Curriculum $curriculum)
     {
+        $this->authorize('delete_curriculums');
         $curriculum->delete();
 
         return redirect()->route('dashboard.curriculums.index')->with('success', 'Curriculum deleted successfully.');
